ex3/requirements.txt
matplotlib==3.10.0
numpy==2.2.0
opencv_python==4.10.0.82
Pillow==11.0.0
